.history/requirements_20220914214253.txt
cohere
python-dotenv